requirements.txt
streamlit==1.39.0
cryptography
pillow
numpy
matplotlib
